enterprise_search clean up: search_index and search_indices
