+++ x-pack/solutions/search/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/constants.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export enum SearchIndexTabId {
+  // all indices
+  OVERVIEW = 'overview',
+  DOCUMENTS = 'documents',
+  INDEX_MAPPINGS = 'index_mappings',
+  PIPELINES = 'pipelines',
+  // connector indices
+  CONFIGURATION = 'configuration',
+  SYNC_RULES = 'sync_rules',
+  SCHEDULING = 'scheduling',
+}
