--- x-pack/solutions/search/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_indices/styles.ts
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { css } from '@emotion/react';
-
-import { type EuiThemeComputed } from '@elastic/eui';
-
-export const entSearchIndicesSearchBar = (euiTheme: EuiThemeComputed<{}>) =>
-  css({
-    maxWidth: `calc(${euiTheme.size.xxl} * 10)`,
-  });
